Add device tools with modular architecture and update existing tools to use genericID schema
